requirements.txt
PyOpenGL>=3.1.10
PyOpenGL-accelerate>=3.1.10
Pillow>=12.0.0
